+++ src/realm/column_table.hpp
@@ -54,7 +54,7 @@ public:
     void adj_acc_move_over(size_t, size_t) noexcept override;
     void adj_acc_clear_root_table() noexcept override;
     void adj_acc_swap_rows(size_t, size_t) noexcept override;
-    void adj_acc_subsume_identity(size_t, size_t) noexcept override;
+    void adj_acc_change_link_targets(size_t, size_t) noexcept override;
     void mark(int) noexcept override;
     void refresh_accessor_tree(size_t, const Spec&) override;
 
@@ -100,7 +100,7 @@ protected:
         template<bool fix_ndx_in_parent>
         void adj_swap_rows(size_t row_ndx_1, size_t row_ndx_2) noexcept;
         template<bool fix_ndx_in_parent>
-        void adj_subsume_identity(size_t row_ndx, size_t subsumed_by) noexcept;
+        void adj_change_link_targets(size_t row_ndx, size_t subsumed_by) noexcept;
 
         void update_accessors(const size_t* col_path_begin, const size_t* col_path_end,
                               _impl::TableFriend::AccessorUpdater&);
@@ -378,10 +378,10 @@ inline void SubtableColumnBase::adj_acc_swap_rows(size_t row_ndx_1, size_t row_n
     m_subtable_map.adj_swap_rows<fix_ndx_in_parent>(row_ndx_1, row_ndx_2);
 }
 
-inline void SubtableColumnBase::adj_acc_subsume_identity(size_t row_ndx, size_t subsumed_by_row_ndx) noexcept
+inline void SubtableColumnBase::adj_acc_change_link_targets(size_t row_ndx, size_t new_row_ndx) noexcept
 {
     const bool fix_ndx_in_parent = true;
-    m_subtable_map.adj_subsume_identity<fix_ndx_in_parent>(row_ndx, subsumed_by_row_ndx);
+    m_subtable_map.adj_change_link_targets<fix_ndx_in_parent>(row_ndx, new_row_ndx);
 }
 
 inline Table* SubtableColumnBase::get_subtable_accessor(size_t row_ndx) const noexcept
@@ -516,9 +516,9 @@ void SubtableColumnBase::SubtableMap::adj_swap_rows(size_t row_ndx_1, size_t row
 }
 
 template<bool fix_ndx_in_parent>
-void SubtableColumnBase::SubtableMap::adj_subsume_identity(size_t row_ndx, size_t subsumed_by_row_ndx) noexcept
+void SubtableColumnBase::SubtableMap::adj_change_link_targets(size_t row_ndx, size_t new_row_ndx) noexcept
 {
-    static_cast<void>(subsumed_by_row_ndx);
+    static_cast<void>(new_row_ndx);
 
     using tf = _impl::TableFriend;
     for (auto it = m_entries.begin(); it != m_entries.end(); ++it) {