+++ src/realm/table.cpp
@@ -2093,15 +2093,15 @@ void Table::erase_row(size_t row_ndx, bool is_move_last_over)
     remove_backlink_broken_rows(state); // Throws
 }
 
-void Table::subsume_identity(size_t row_ndx, size_t subsumed_by_row_ndx)
+void Table::change_link_targets(size_t row_ndx, size_t new_row_ndx)
 {
     REALM_ASSERT(is_attached());
     REALM_ASSERT_EX(row_ndx < m_size, row_ndx, m_size);
 
-    do_subsume_identity(row_ndx, subsumed_by_row_ndx);
+    do_change_link_targets(row_ndx, new_row_ndx);
 
     if (Replication* repl = get_repl()) {
-        repl->subsume_identity(this, row_ndx, subsumed_by_row_ndx);
+        repl->change_link_targets(this, row_ndx, new_row_ndx);
     }
 }
 
@@ -2233,12 +2233,12 @@ void Table::do_swap_rows(size_t row_ndx_1, size_t row_ndx_2)
 }
 
 
-void Table::do_subsume_identity(size_t row_ndx, size_t subsumed_by_row_ndx)
+void Table::do_change_link_targets(size_t row_ndx, size_t new_row_ndx)
 {
     // Replace links through backlink columns, WITHOUT generating SetLink instructions.
     //
     // This bypasses handling of cascading rows, and we have decided that this is OK, because
-    // SubsumeIdentity is always followed by MoveLastOver, so breaking the last strong link
+    // ChangeLinkTargets is always followed by MoveLastOver, so breaking the last strong link
     // to a row that is being subsumed will have no observable effect, while honoring the
     // cascading behavior would complicate the calling code somewhat (having to take
     // into account whether or not the row was removed as a consequence of cascade, leading
@@ -2259,21 +2259,21 @@ void Table::do_subsume_identity(size_t row_ndx, size_t subsumed_by_row_ndx)
             size_t origin_row_ndx = col.get_backlink(row_ndx, 0);
 
             if (origin_col_type == col_type_Link) {
-                origin_table.do_set_link(origin_col_ndx, origin_row_ndx, subsumed_by_row_ndx);
+                origin_table.do_set_link(origin_col_ndx, origin_row_ndx, new_row_ndx);
             }
             else if (origin_col_type == col_type_LinkList) {
                 LinkViewRef links = origin_table.get_linklist(origin_col_ndx, origin_row_ndx);
                 for (size_t j = 0; j < links->size(); ++j) {
                     using llf = _impl::LinkListFriend;
                     if (links->get(j).get_index() == row_ndx) {
-                        llf::do_set(*links, j, subsumed_by_row_ndx);
+                        llf::do_set(*links, j, new_row_ndx);
                     }
                 }
             }
         }
     }
 
-    adj_acc_subsume_identity(row_ndx, subsumed_by_row_ndx);
+    adj_acc_change_link_targets(row_ndx, new_row_ndx);
     bump_version();
 }
 
@@ -5004,19 +5004,19 @@ void Table::adj_acc_swap_rows(size_t row_ndx_1, size_t row_ndx_2) noexcept
 }
 
 
-void Table::adj_acc_subsume_identity(size_t row_ndx, size_t subsumed_by_row_ndx) noexcept
+void Table::adj_acc_change_link_targets(size_t row_ndx, size_t new_row_ndx) noexcept
 {
     // This function must assume no more than minimal consistency of the
     // accessor hierarchy. This means in particular that it cannot access the
     // underlying node structure. See AccessorConsistencyLevels.
 
-    adj_row_acc_subsume_identity(row_ndx, subsumed_by_row_ndx);
+    adj_row_acc_change_link_targets(row_ndx, new_row_ndx);
 
     // Adjust subtable/linklist/mixed accessors after subsume identity
     size_t n = m_cols.size();
     for (size_t i = 0; i < n; ++i) {
         if (ColumnBase* col = m_cols[i])
-            col->adj_acc_subsume_identity(row_ndx, subsumed_by_row_ndx);
+            col->adj_acc_change_link_targets(row_ndx, new_row_ndx);
     }
 }
 
@@ -5138,13 +5138,13 @@ void Table::adj_row_acc_swap_rows(size_t row_ndx_1, size_t row_ndx_2) noexcept
 }
 
 
-void Table::adj_row_acc_subsume_identity(size_t row_ndx, size_t subsumed_by_row_ndx) noexcept
+void Table::adj_row_acc_change_link_targets(size_t row_ndx, size_t new_row_ndx) noexcept
 {
     // This function must assume no more than minimal consistency of the
     // accessor hierarchy. This means in particular that it cannot access the
     // underlying node structure. See AccessorConsistencyLevels.
 
-    static_cast<void>(subsumed_by_row_ndx);
+    static_cast<void>(new_row_ndx);
 
     LockGuard lock(m_accessor_mutex);
     RowBase* row = m_row_accessors;